move circuit to own module
